pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "build"]
build-backend = "setuptools.build_meta"

[project]
name = "mifarepy"
version = "2.0"
description = "Python library for interfacing with MIFARE RFID card readers using mifarepy protocol"
authors = [
    { name = "Spark Drago", email = "huzaifa.farooq05@gmail.com" }
]
license = { file = "LICENSE" }
dependencies = [
    "pyserial"
]
requires-python = ">=3.6"
readme = { file = "README.md", content-type = "text/markdown" }
classifiers = [
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent"
]

[tool.setuptools]
packages = ["mifarepy"]

[tool.setuptools.package-data]
mifarepy = ["*.txt", "*.md"]

[tool.setuptools.dynamic]
version = {attr = "mifarepy.__version__"}
